+++ src/utils/session.ts
@@ -1,26 +1,30 @@
 import type { CookieSerializeOptions } from "cookie-es";
 import crypto from "uncrypto";
-import {
-  seal,
-  unseal,
-  defaults as sealDefaults,
-} from "../../lib/iron-webcrypto";
-import type { SealOptions } from "../../lib/iron-webcrypto";
+import { seal, unseal, defaults as sealDefaults } from "iron-webcrypto";
+import type { SealOptions } from "iron-webcrypto";
 import type { H3Event } from "../event";
 import { getCookie, setCookie } from "./cookie";
 
-type SessionDataT = Record<string, string | number | boolean>;
+type SessionDataT = Record<string, any>;
 export type SessionData<T extends SessionDataT = SessionDataT> = T;
 
 export interface Session<T extends SessionDataT = SessionDataT> {
   id: string;
+  createdAt: number;
   data: SessionData<T>;
 }
 
 export interface SessionConfig {
+  /** Private key used to encrypt session tokens */
   password: string;
+  /** Session expiration time in seconds */
+  maxAge?: number;
+  /** default is h3 */
   name?: string;
-  cookie?: CookieSerializeOptions;
+  /** Default is secure, httpOnly, / */
+  cookie?: false | CookieSerializeOptions;
+  /** Default is x-h3-session / x-{name}-session */
+  sessionHeader?: false | string;
   seal?: SealOptions;
   crypto?: Crypto;
 }
@@ -73,24 +77,43 @@ export async function getSession<T extends SessionDataT = SessionDataT>(
   }
 
   // Prepare an empty session object and store in context
-  const session: Session<T> = { id: "", data: Object.create(null) };
+  const session: Session<T> = {
+    id: "",
+    createdAt: 0,
+    data: Object.create(null),
+  };
   event.context.sessions![sessionName] = session;
 
-  // Try to hydrate from cookies
-  const reqCookie = getCookie(event, sessionName);
-  if (!reqCookie) {
-    // New session store in response cookies
-    session.id = (config.crypto || crypto).randomUUID();
-    await updateSession(event, config);
-  } else {
+  // Try to load session
+  let sealedSession: string | undefined;
+  // Try header first
+  if (config.sessionHeader !== false) {
+    const headerName =
+      typeof config.sessionHeader === "string"
+        ? config.sessionHeader.toLowerCase()
+        : `x-${sessionName.toLowerCase()}-session`;
+    const headerValue = event.node.req.headers[headerName];
+    if (typeof headerValue === "string") {
+      sealedSession = headerValue;
+    }
+  }
+  // Fallback to cookies
+  if (!sealedSession) {
+    sealedSession = getCookie(event, sessionName);
+  }
+  if (sealedSession) {
     // Unseal session data from cookie
-    const unsealed = await unseal(
-      config.crypto || crypto,
-      reqCookie,
-      config.password,
-      config.seal || sealDefaults
+    const unsealed = await unsealSession(event, config, sealedSession).catch(
+      () => {}
     );
     Object.assign(session, unsealed);
+  }
+
+  // New session store in response cookies
+  if (!session.id) {
+    session.id = (config.crypto || crypto).randomUUID();
+    session.createdAt = Date.now();
+    await updateSession(event, config);
   }
 
   return session;
@@ -121,15 +144,62 @@ export async function updateSession<T extends SessionDataT = SessionDataT>(
   }
 
   // Seal and store in cookie
-  const sealed = await seal(
-    config.crypto || crypto,
-    session,
-    config.password,
-    config.seal || sealDefaults
-  );
-  setCookie(event, sessionName, sealed, config.cookie || DEFAULT_COOKIE);
+  if (config.cookie !== false) {
+    const sealed = await sealSession(event, config);
+    setCookie(event, sessionName, sealed, {
+      ...DEFAULT_COOKIE,
+      expires: config.maxAge
+        ? new Date(session.createdAt + config.maxAge * 1000)
+        : undefined,
+      ...config.cookie,
+    });
+  }
 
   return session;
+}
+
+export async function sealSession<T extends SessionDataT = SessionDataT>(
+  event: H3Event,
+  config: SessionConfig
+) {
+  const sessionName = config.name || DEFAULT_NAME;
+
+  // Access current session
+  const session: Session<T> =
+    (event.context.sessions?.[sessionName] as Session<T>) ||
+    (await getSession<T>(event, config));
+
+  const sealed = await seal(config.crypto || crypto, session, config.password, {
+    ...sealDefaults,
+    ttl: config.maxAge ? config.maxAge * 1000 : 0,
+    ...config.seal,
+  });
+
+  return sealed;
+}
+
+export async function unsealSession(
+  _event: H3Event,
+  config: SessionConfig,
+  sealed: string
+) {
+  const unsealed = (await unseal(
+    config.crypto || crypto,
+    sealed,
+    config.password,
+    {
+      ...sealDefaults,
+      ttl: config.maxAge ? config.maxAge * 1000 : 0,
+      ...config.seal,
+    }
+  )) as Partial<Session>;
+  if (config.maxAge) {
+    const age = Date.now() - (unsealed.createdAt || Number.NEGATIVE_INFINITY);
+    if (age > config.maxAge * 1000) {
+      throw new Error("Session expired!");
+    }
+  }
+  return unsealed;
 }
 
 export async function clearSession(event: H3Event, config: SessionConfig) {
@@ -137,5 +207,8 @@ export async function clearSession(event: H3Event, config: SessionConfig) {
   if (event.context.sessions?.[sessionName]) {
     delete event.context.sessions![sessionName];
   }
-  await setCookie(event, sessionName, "", config.cookie || DEFAULT_COOKIE);
+  await setCookie(event, sessionName, "", {
+    ...DEFAULT_COOKIE,
+    ...config.cookie,
+  });
 }